+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on [Keep a Changelog], and this project adheres to
 
 - Added `Command`, `Event` and `Timeout` as aliases for `Message` in preparation for string message typing
 - **[APP BC]** Added `MessageDescription()` method to `Message` interface
+- **[APP BC]** Added `Validate()` method to `Message` interface
 - **[ENGINE BC]** Added `ProjectionConfigurer.DeliveryPolicy()`
 - **[ENGINE BC]** Added `ProjectionScope.IsPrimaryDelivery()`
 
@@ -23,10 +24,12 @@ The format is based on [Keep a Changelog], and this project adheres to
 
 - **[APP BC]** Handler and application identity keys must now be an RFC 4122 UUID string
 
-### Deprecated
+### Removed
 
-- Deprecated `DescribableMessage` interface, as all messages are now describable
-- Deprecated `DescribeMessage()` function
+- **[ENGINE BC]** Removed `DescribableMessage` interface, as all messages are now describable
+- **[ENGINE BC]** Removed `ValidatableMessage` interface, as all messages are now validatable
+- **[ENGINE BC]** Removed `DescribeMessage()` function
+- **[ENGINE BC]** Removed `ValidateMessage()` function
 
 ## [0.11.1] - 2021-03-01